Fix opener-coupled link classification and split wrap tests (#293)
Stop passing opening-punct-stripped text to looks_like_link so bracket-
coupled links classify correctly, add Markdown link regression cases, and
split src/wrap/tests.rs into focused modules under the 400-line limit.
